shifting missing test
