+++ crates/orderbook/src/arguments.rs
@@ -233,49 +233,3 @@ impl std::fmt::Display for Arguments {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use maplit::hashmap;
-
-    #[test]
-    fn parse_partner_fee_factor_ok() {
-        let x = "0x0000000000000000000000000000000000000000000000000000000000000000";
-        let y = "0x0101010101010101010101010101010101010101010101010101010101010101";
-        // without spaces
-        assert_eq!(
-            parse_partner_fee_factor(&format!("{}:0.5,{}:0.7", x, y)).unwrap(),
-            hashmap! { AppId([0u8; 32]) => 0.5, AppId([1u8; 32]) => 0.7 }
-        );
-        // with spaces
-        assert_eq!(
-            parse_partner_fee_factor(&format!("{}: 0.5, {}: 0.7", x, y)).unwrap(),
-            hashmap! { AppId([0u8; 32]) => 0.5, AppId([1u8; 32]) => 0.7 }
-        );
-        // whole numbers
-        assert_eq!(
-            parse_partner_fee_factor(&format!("{}: 1, {}: 2", x, y)).unwrap(),
-            hashmap! { AppId([0u8; 32]) => 1., AppId([1u8; 32]) => 2. }
-        );
-    }
-
-    #[test]
-    fn parse_partner_fee_factor_err() {
-        assert!(parse_partner_fee_factor("0x1:0.5,0x2:0.7").is_err());
-        assert!(parse_partner_fee_factor("0x12:0.5,0x22:0.7").is_err());
-        assert!(parse_partner_fee_factor(
-            "0x0000000000000000000000000000000000000000000000000000000000000000:0.5:3"
-        )
-        .is_err());
-        assert!(parse_partner_fee_factor(
-            "0x0000000000000000000000000000000000000000000000000000000000000000:word"
-        )
-        .is_err());
-    }
-
-    #[test]
-    fn parse_partner_fee_factor_ok_on_empty() {
-        assert!(parse_partner_fee_factor("").unwrap().is_empty());
-    }
-}
